feat: 프리셋 화면에서 새로운 프리셋 추가할 수 있도록 함

+++ app/src/main/java/com/example/android/waterwise/data/hydratepreset/HydratePreset.kt
@@ -15,8 +15,8 @@ import com.example.android.waterwise.data.beverage.Beverage
     )]
 )
 data class HydratePreset(
-    @PrimaryKey(autoGenerate = true) @ColumnInfo(name = "hydrate_preset_id") val id: Int = 0,
+    @PrimaryKey(autoGenerate = true) @ColumnInfo(name = "hydrate_preset_id") val id: Long = 0,
     @ColumnInfo(name = "amount") val amount: Int,
-    @ColumnInfo(name = "beverage_id") val beverageId: Int,
+    @ColumnInfo(name = "beverage_id") val beverageId: Long,
     @ColumnInfo(name = "nickname") val nickname: String,
 )
